Replace admin user related atoms using jotai
