+++ src/hooks/admin/useAdminUser.tsx
@@ -1,17 +1,15 @@
 import useApi from '@hooks/useApi';
 import { User, Workspace } from '@@types/index';
-import { useSetRecoilState } from 'recoil';
-import { adminUserAtom } from '@recoils/atoms';
 import { useNavigate } from 'react-router-dom';
 import useAuthentication from '@hooks/useAuthentication';
 import { useSetAtom } from 'jotai';
-import { banksAtom, workspacesAtom } from 'src/jotai/admin/atoms';
+import { adminUserAtom, banksAtom, workspacesAtom } from 'src/jotai/admin/atoms';
 
 function useAdminUser() {
   const { adminApi } = useApi();
   const { logout } = useAuthentication();
   const setWorkspaces = useSetAtom(workspacesAtom);
-  const setAdminUser = useSetRecoilState(adminUserAtom);
+  const setAdminUser = useSetAtom(adminUserAtom);
   const setBanks = useSetAtom(banksAtom);
   const navigate = useNavigate();
